renamed namespace to Sequence.Adapter
