+++ Assets/SequenceSDK/Adapter.meta
