+++ Assets/SequenceSDK/Adapter/Tests/SequenceAdapterTests.asmdef
