+++ Packages/Sequence-Unity/Sequence/SequenceSDK/Adapter.meta
