+++ Packages/Sequence-Unity/Sequence/SequenceSDK/Adapter/ISequence.cs.meta
