+++ Packages/Sequence-Unity/Sequence/SequenceSDK/Adapter/SequenceAdapter.asmdef
@@ -1,6 +1,6 @@
 {
-    "name": "SequenceUnified",
-    "rootNamespace": "Sequence.Unified",
+    "name": "SequenceAdapter",
+    "rootNamespace": "Sequence.Adapter",
     "references": [
         "GUID:a35e3a53d4439435f8b36ed2c6158cd8",
         "GUID:4e0a798abbda240658187632ae443a67",
